requirements.txt
pandas
numpy
yfinance
statsmodels
matplotlib
pyarrow
PyYAML
